requirements.txt
Django==5.1.3
psycopg2-binary==2.9.10
python-decouple==3.8
PyPDF2==3.0.1
python-pptx==1.0.2